requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.2.1
imap-tools==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
flask==2.3.3
werkzeug==2.3.7
# Added for Streamlit app
streamlit==1.44.1
pandas==2.0.3
plotly==6.0.1
requests==2.31.0